Liquid physics: respect flow direction
Generalize existing logic handling flowing down neighbors and also apply
it to directional source nodes. This adds the logic to decode direction
from directional param2.

Directional flows with a direction (including flowing down) will only
spread in that direction.

Note that when placing a directional source with param2 == 0 above air it
requires one update cycle for the source to update itself to the flowing
down state, so neighboring nodes will get flooded during this first cycle.
To prevent this, set param2 to 248 when placing the source (which is also
safe if not placed above air, it will then take one cycle to update to the
non flowing down state).