Liquid physics: refactor source neighbor handling
The lower neighbor is special cased and a case statement for the other neighbor
types is added.  This provides a frame where new physics logic can be added
easily.

Note that even though this change prevents the selection of liquid_kind from a
lower source node, this is no semantic change, because the lower neighbor is
checked last and in that case there were no other neighboring sources or flows
leading to no liquid entering the current node. This change makes that fact
more explicit and could be considered a fix.