Liquid physics: directional liquids flow towards closest flow down
A directional liquid that can flow down will become directional in the down
direction.

A directional liquid that cannot flow down checks whether there are any
neighbors with a direction. It will choose to flow into the first neighbor
whose distance component indicates the shortest path to a flow down if the
resulting path is shorter than liquid_directed_range.

To allow neighboring directional source nodes to pick a direction, too, these
need to be pushed to the update queue on change.