Liquid visualization: give volume to linear flows
Flow corners surrounded by air nodes are forced to the lower edge of the node
to visualize a flow turning into air.

For directional flows the number of air nodes required is increased from 2 to 3
to prevent linear flows being rendered flat.

Nother that this still leaves the rendering of directional flows somewhat
suboptimal.

This could also applied to non directional flows, because it only has a very
subtle effect on the outer edges, but the change is restricted to directional
liquids to not introduce any change in existing worlds.

TODO: decide whether to apply this change also to non directional flows