Liquid visualization: fix waterfalls bleeding into neighboring flows
A flow flowing down is visualized as flowing into any neighboring flow of the
same kind even though in the actual physics implementation it only flows down.

This commit changes the visualization of directional liquids to flow into the
waterfall.

Unfortunately the solution in this commit is probably not simple enough for
this minor visual error.

TODO: decide whether this fix is worth it and perhaps apply it to all liquids